+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 ### Enhancements
 - [CM-451] Add support for setting a prefilled message to send before launching a chat.
 - [CM-356] Add support for language change rich message.
+- [CM-474] Add send message API for sending text and rich messages.
 
 ## [5.7.0] - 2020-09-10